Fix cases where the type of a for loop variable is smaller than the type of the "end" variable. In some cases the usage may have been benign since it can be seen by code inspection that the maximum value of the end variable can't exceed the maximum value of the loop variable. However, on general principles, all the usages have been fixed.
svn path=/trunk/; revision=33692

+++ epan/dissectors/packet-iec104.c
@@ -1336,10 +1336,10 @@ static void dissect_iec104asdu(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tr
 /* Is is called twice: For 'Packet List' and for 'Packet Details' */
 static void dissect_iec104apci(tvbuff_t *tvb, packet_info *pinfo, proto_tree *tree)
 {
-	guint8 TcpLen = tvb_length(tvb);
-	guint16 Brossa = 0;
+	guint TcpLen = tvb_length(tvb);
+	guint Brossa = 0;
 	guint8 Start;
-	guint8 Off;
+	guint Off;
 	guint8 Byte1 = 0;
 	struct apciheader * apcih;
 	emem_strbuf_t * res;
